--- apps/web/src/App.tsx
@@ -1,9 +1,0 @@
-import { Stack } from "@/lib/stackflow";
-import "@stackflow/plugin-basic-ui/index.css";
-
-const App = () => {
-
-  return <Stack />;
-}
-
-export default App
