+++ apps/web/src/app/AppRouter.tsx
@@ -1,0 +1,33 @@
+import { RouterProvider, createBrowserRouter } from 'react-router-dom'
+import App from './App'
+import HomePage from '@/pages/homePage'
+import LiveRecordPage from '@/pages/liveRecordPage'
+import LoginPage from '@/pages/auth/ui/loginPage'
+import TeamSelectPage from '@/pages/auth/ui/teamSelectPage'
+import NickNamePage from '@/pages/auth/ui/nickNamePage'
+
+import { PATH } from '@/constants/path'
+
+const AppRouter = () => {
+  const router = createBrowserRouter([
+    {
+      path: PATH.HOME,
+      element: <App />,
+      children: [
+        { path: '', element: <HomePage /> },
+        { path: PATH.LIVE_RECORD, element: <LiveRecordPage /> },
+        { path: PATH.LOGIN, element: <LoginPage /> },
+        { path: PATH.TEAM_SELECT, element: <TeamSelectPage /> },
+        { path: PATH.NICKNAME, element: <NickNamePage /> },
+        // {
+        //   path: 'test/:id',
+        //   element: <TestId />,
+        // },
+      ],
+    },
+  ])
+
+  return <RouterProvider router={router} />
+}
+
+export default AppRouter
