+++ apps/web/src/entities/auth/ui/index.ts
@@ -1,0 +1,2 @@
+export * from './Banner'
+export * from './TeamButton'
