+++ apps/web/src/entities/game/index.ts
@@ -1,0 +1,1 @@
+export * from './ui'
