+++ apps/web/src/entities/game/ui/index.ts
@@ -1,0 +1,1 @@
+export * from './gameCard'
