+++ apps/web/src/features/auth/index.ts
@@ -1,0 +1,1 @@
+export * from './ui'
