+++ apps/web/src/pages/auth/ui/index.ts
@@ -1,0 +1,3 @@
+export * from './loginPage'
+export * from './teamSelectPage'
+export * from './nickNamePage'
