+++ apps/web/src/pages/homePage.tsx
@@ -1,12 +1,11 @@
 import { AppScreen } from '@stackflow/plugin-basic-ui'
-import { useFlow } from '@/lib/stackflow'
-import { Button } from '@/components/ui/button'
-import GlobalNavigationBar from '@/components/common/globalNavigationBar'
-import Header from '@/components/common/Header'
+import { useFlow } from '@/share/lib/stackflow'
+import { Button } from '@/share/ui/common/Button'
+import { GlobalNavigationBar } from '@/widget/navigation'
+import { Header } from '@/widget/header'
 import DarkBallogLogo from '@/assets/darkBallogLogo.svg'
 import GameCardCarousel from '@/features/game/components/gameCardCarousel'
 import type { ActivityComponentType } from '@stackflow/react'
-
 
 const HomePage: ActivityComponentType = () => {
   const { replace } = useFlow()
@@ -14,7 +13,9 @@ const HomePage: ActivityComponentType = () => {
   return (
     <>
       <AppScreen>
-        <Header logo={<img src={DarkBallogLogo} alt="ballog" className="h-6" />} />
+        <Header
+          logo={<img src={DarkBallogLogo} alt="ballog" className="h-6" />}
+        />
         <h2 className="text-center mt-15 mb-8 text-xl font-bold">
           오늘 경기 선택하고 감정 기록하기
         </h2>
@@ -29,10 +30,7 @@ const HomePage: ActivityComponentType = () => {
           </p>
         </div>
         <div className="flex justify-center mt-10">
-          <Button
-            className="w-30 h-10"
-            onClick={() => replace('Login', {})}
-          >
+          <Button className="w-30 h-10" onClick={() => replace('Login', {})}>
             기록 시작하기
           </Button>
         </div>
