+++ apps/web/src/share/ui/common/Button/index.ts
@@ -1,0 +1,1 @@
+export * from './button'
