+++ apps/web/src/stories/Button.stories.ts
@@ -3,7 +3,7 @@ import type { Meta, StoryObj } from '@storybook/react-vite'
 
 import { fn } from 'storybook/test'
 
-import { Button } from '@/components/ui/button'
+import { Button } from '@/share/ui/common/Button/button'
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 const meta = {
